python/requirements.txt
# ============================================================================
# SIGHT ONE (瞰析 ONE) - Python Environment Requirements
# Intelligent Vision Analysis Platform
# ============================================================================

# ----------------------------------------------------------------------------
# Core Scientific Computing & Image Processing
# ----------------------------------------------------------------------------
numpy>=1.24.0,<2.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
scipy>=1.7.0

# ----------------------------------------------------------------------------
# Deep Learning & Computer Vision
# ----------------------------------------------------------------------------
torch>=2.0.0
ultralytics>=8.0.0  # YOLOv8 for object detection

# ----------------------------------------------------------------------------
# Drone Control
# ----------------------------------------------------------------------------
djitellopy>=2.5.0

# ----------------------------------------------------------------------------
# QR Code Detection
# ----------------------------------------------------------------------------
pyzbar>=0.1.9

# ----------------------------------------------------------------------------
# AI Services & LLM Integration
# ----------------------------------------------------------------------------
# OpenAI (standard and Azure)
openai>=1.30.0

# Anthropic Claude
anthropic>=0.25.0

# Google Gemini
google-generativeai>=0.3.0

# Azure Services (optional)
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0

# Alibaba Cloud AI (optional)
dashscope>=1.10.0

# ----------------------------------------------------------------------------
# Async Web Services & Networking
# ----------------------------------------------------------------------------
websockets>=11.0.0
aiohttp>=3.8.0
httpx>=0.24.0  # Modern async HTTP client
requests>=2.31.0

# ----------------------------------------------------------------------------
# Data Processing & Validation
# ----------------------------------------------------------------------------
pandas>=1.3.0
pydantic[email]>=2.5.0
pydantic-settings>=2.0.0
jsonschema>=4.0.0
python-dateutil>=2.8.0

# ----------------------------------------------------------------------------
# Configuration & Environment
# ----------------------------------------------------------------------------
python-dotenv>=1.0.0
click>=8.1.0

# ----------------------------------------------------------------------------
# Logging & Monitoring
# ----------------------------------------------------------------------------
loguru>=0.6.0

# ----------------------------------------------------------------------------
# System Utilities
# ----------------------------------------------------------------------------
psutil>=5.8.0
pyserial>=3.5  # Serial communication (if needed)

# ----------------------------------------------------------------------------
# Development & Testing
# ----------------------------------------------------------------------------
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
pytest-cov>=4.0.0

# ----------------------------------------------------------------------------
# FastAPI & Web Framework (for API services)
# ----------------------------------------------------------------------------
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6  # For file uploads

# ----------------------------------------------------------------------------
# Database & Storage (optional, for future use)
# ----------------------------------------------------------------------------
# sqlalchemy>=2.0.0
# alembic>=1.12.0
# redis>=5.0.0

# ----------------------------------------------------------------------------
# Image Processing Enhancement
# ----------------------------------------------------------------------------
scikit-image>=0.21.0  # Advanced image processing
albumentations>=1.3.0  # Image augmentation (optional)

# ----------------------------------------------------------------------------
# Message Queue & Task Processing (optional)
# ----------------------------------------------------------------------------
# celery>=5.3.0
# kombu>=5.3.0

# ----------------------------------------------------------------------------
# Documentation
# ----------------------------------------------------------------------------
# sphinx>=7.0.0
# sphinx-rtd-theme>=1.3.0

# ----------------------------------------------------------------------------
# Security & Authentication (for production)
# ----------------------------------------------------------------------------
# python-jose[cryptography]>=3.3.0
# passlib[bcrypt]>=1.7.4
# python-multipart>=0.0.6

# ----------------------------------------------------------------------------
# Monitoring & Observability
# ----------------------------------------------------------------------------
# prometheus-client>=0.18.0
# sentry-sdk>=1.38.0

# ----------------------------------------------------------------------------
# Additional Utilities
# ----------------------------------------------------------------------------
tqdm>=4.66.0  # Progress bars
colorama>=0.4.6  # Colored terminal output
rich>=13.0.0  # Rich terminal formatting
tenacity>=8.2.0  # Retry logic
cachetools>=5.3.0  # Caching utilities

# ============================================================================
# NOTES:
# ============================================================================
# 1. Python Version: Requires Python 3.9 or higher
# 2. CUDA Support: For GPU acceleration, install PyTorch with CUDA:
#    pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
# 3. Windows Users: Some packages may require Visual C++ Build Tools
# 4. WSL/Linux: Ensure libGL is installed for OpenCV:
#    sudo apt-get install libgl1-mesa-glx
# 5. Development: Install all dependencies with:
#    pip install -r requirements.txt
# 6. Production: Comment out development dependencies for lighter deployment
# ============================================================================